requirements.txt
flask
flask-cors
pdfplumber
python-docx
spacy
werkzeug
flasgger